refactor: change from Writable::buildFromReadable to Readable->cloneIntoWritable
